ROCANA-3570: Impala pulls in jars that become broken symlinks
Our HDP version of Impala was pulling in two extra jars that the Cloudera version did not pull in. These came in from HDP versions of components we were using. The problem is the cdh-package project that creates our Impala RPM (for HDP) found these two jars and generated broken symlinks for them. This commit excludes those two jars so we won't have the broken symlinks anymore.

I did a `mvn clean package` to test and it still builds fine.

I used `mvn dependency:tree` to see how the jars were pulled in and to make sure they were totally gone afterwards. After this commit the only thing that changed in the dependency tree is those two jars were missing, which is exactly what we wanted.

Here are the full coordinates of the jars removed:
 - org.apache.zookeeper:zookeeper:test-jar:tests:3.4.6.2.3.0.0-2557:compile
 - org.apache.hive.shims:hive-shims-0.20S:jar:1.2.1.2.3.0.0-2557:runtime