ROCANA-4170: Impala does not compile on CentOS 7 due to Berkeley DB 5.x
The options to fix are:
1. Change the preprocessor code so it treats 5.0 the same as 4.1. The
current code says "if 4.1, 4.2, 4.3 etc. then use the new function
definition, else use the old one", but that doesn't work when the major
version increases. We can modify the code so it says "if 4.1, 4.2, 4.3
etc. or >= 5 then use the new function definition".
2. Use Berkeley DB 4, which also comes on CentOS 7 (see
/usr/include/libdb4/db.h, which is version 4.8.30). We'd have to change
the Impala build process to compile/link against Berkeley DB 4

I went with #1 in this commit. #2 seemed like a harder fix. Plus, I
wasn't sure what would happen if we ran it on a CentOS 7 machine that
had Berkeley DB 5 as the default.